backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Core dependencies (inherited from parent)
groq>=0.9.0
supabase>=2.1.0
python-dotenv>=1.0.0
websockets>=11.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI endpoints
